feat: remove use of moto dynamodb2 from tests
Moves the local DynamoDB run to the package level and all tests use
it instead of running moto's incomplete dynamodb mock.

This also fixes a bug in the tests that left autopush.db.key_hash
set after an endpoint test ran.
