+++ autopush/tests/test_web_webpush.py
@@ -3,12 +3,11 @@ import uuid
 
 from cryptography.fernet import Fernet
 from mock import Mock
-from moto import mock_dynamodb2
 from nose.tools import eq_, ok_
 from twisted.internet.defer import inlineCallbacks
 from twisted.trial import unittest
 
-from autopush.db import Message, create_rotating_message_table
+from autopush.db import Message
 from autopush.http import EndpointHTTPFactory
 from autopush.router.interface import IRouter, RouterResponse
 from autopush.settings import AutopushSettings
@@ -18,16 +17,6 @@ from autopush.tests.support import test_db
 dummy_uaid = str(uuid.UUID("abad1dea00000000aabbccdd00000000"))
 dummy_chid = str(uuid.UUID("deadbeef00000000decafbad00000000"))
 dummy_token = dummy_uaid + ":" + dummy_chid
-mock_dynamodb2 = mock_dynamodb2()
-
-
-def setUp():
-    mock_dynamodb2.start()
-    create_rotating_message_table()
-
-
-def tearDown():
-    mock_dynamodb2.stop()
 
 
 class TestWebpushHandler(unittest.TestCase):
